+++ CHANGELOG.md
@@ -5,15 +5,24 @@ All notable changes to this project will be documented in this file.
 ## Next version
 
 ### Features
+- [Core] Change /find-player to /find #417
+- [Core] Using `/regular <name>` now defaults to promoting the player. #417
+- [Diggy] Added new formula to calculate experience requirement for next level #402
+- [Diggy] Added health bonus on level up #402
+- [Diggy] Added new level system to replace stone sent to surface #402
+- [Diggy] Mining rocks no longer gives stone or coal #402
 - [Diggy] Added particles when a biter mines a rock upon spawning #424
 - [Diggy] Added bot mining #434
 - [Map] Add Tetris map #433
 - [Map] Add World Thanksgiving map #433
+- [Diggy] Added bot mining levels and rock mining/destruction particles #442
 
 ### Bugfixes
 - [Diggy] Stones killed by damage no longer spill. #395
 - [Core] Fix /kill non-functional if walkabout is disabled. Fix walkabout giving from variable definition. #425
+- [Diggy] Improved biter spawning algorithm #408
 - [Core] Fix null reference in chat_triggers #431
+- [Core] Fix nil ref in train_station_names #441
 
 ### Internal
 - [Core] Cleanup of code formatting. #413 #415 #414 #412 #411